Fix another bug when setting the src file!!! I believe that this must be discussed among the developers.

+++ apps/shallow-sphere/mapc2m.f
@@ -10,12 +10,12 @@ c     Points in [-1,1]x[-1,1]  are mapped to the upper hemisphere
 c     Points in [-3,-1]x[-1,1] are mapped to the lower hemisphere
 c
       implicit real*8(a-h,o-z)
-cf2py integer intent(in) x1
-cf2py integer intent(in) y1
-cf2py integer intent(in) xp
-cf2py integer intent(in) yp
-cf2py integer intent(in) zp
-cf2py integer intent(in) Rsphere
+cf2py  double precision intent(in) x1
+cf2py  double precision intent(in) y1
+cf2py  double precision intent(in) xp
+cf2py  double precision intent(in) yp
+cf2py  double precision intent(in) zp
+cf2py  double precision intent(in) Rsphere
 
 
       r1 = Rsphere
